lesson3/seminar/step1_dvc_mlflow_pipeline/pyproject.toml
[tool.poetry]
name = "lesson3-step1"
version = "0.1.0"
description = "ML Pipeline orchestration with DVC and MLflow"
authors = ["MLOps Course"]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.0"
scikit-learn = "^1.3.0"
mlflow = "^2.8.0"
dvc = {extras = ["s3"], version = "^3.0.0"}
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
black = "^23.11.0"
flake8 = "^6.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88
